Generic updateZone on Template creation and update
