+++ src/Synapse/Cmf/Framework/Theme/Template/Domain/Command/AbstractCommand.php
@@ -7,6 +7,7 @@ use Majora\Framework\Domain\Action\Dal\DalActionTrait;
 use Synapse\Cmf\Framework\Theme\Template\Entity\Template;
 use Synapse\Cmf\Framework\Theme\Template\Model\TemplateInterface;
 use Synapse\Cmf\Framework\Theme\Zone\Entity\ZoneCollection;
+use Synapse\Cmf\Framework\Theme\Zone\Domain\DomainInterface as ZoneDomain;
 
 /**
  * Base class for Template commands.
@@ -24,6 +25,32 @@ abstract class AbstractCommand implements ActionInterface
      * @var ZoneCollection
      */
     protected $zones;
+
+    /**
+     * @var ZoneDomain
+     */
+    protected $zoneDomain;
+
+    /**
+     * Template creation event handler.
+     *
+     * @param TemplateEvent $event
+     */
+    public function onTemplateCreated(TemplateEvent $event)
+    {
+        $template = $event->getTemplate();
+        $zones = $template->getZones();
+        foreach ($template->getTemplateType()->getZoneTypes() as $zoneType) {
+            if ($zones->search(array('zoneType' => $zoneType))->isEmpty()) {
+                if (null === $this->zoneDomain) {
+                    throw new \DomainException('No zoneDomain was set, you are unable to update your configuration.');
+                }
+
+                $zones->add($this->zoneDomain->create($zoneType));
+            }
+        }
+        $template->setZones($zones);
+    }
 
     /**
      * Return related Template if defined.
@@ -57,5 +84,15 @@ abstract class AbstractCommand implements ActionInterface
     public function getZones()
     {
         return $this->zones;
+    }
+
+    /**
+     * Set zoneDomain.
+     *
+     * @param ZoneDomain $zoneDomain
+     */
+    public function setZoneDomain(ZoneDomain $zoneDomain)
+    {
+        $this->zoneDomain = $zoneDomain;
     }
 }
